IT WORKS \o/ 1mil hands in 6 seconds atm

+++ include/Strategy.h
@@ -6,7 +6,7 @@ class Strategy
 public:
     Strategy(){};
 
-    double GetNumUnits(int primaryCount);
+    double GetNumUnits(double primaryCount);
 
 private:
